Decoder: bikeSpeedMs as Float; thresholds at exact km/h equivalents
bikeSpeedMs is Float, carrying the decoder's native 0.25 m/s
resolution. Threshold values picked so each gate fires on the
same raw radar bytes as the prior km/h thresholds:
ALONGSIDE_RIDER_SLOW_MS = 2.75 (raw 0..11),
stationaryMsThreshold = 0.5 (raw 0..2),
riderSpeedFloorMs = 4.25 (raw 17+),
urban-cut at 8.25 (raw 0..33).

Settings slider, prefs storage, and the HA close-pass MQTT
payload stay in km/h via boundary conversion.

+++ app/src/main/java/es/jjrh/bikeradar/RadarState.kt
@@ -65,11 +65,12 @@ data class RadarState(
     val scenarioTimeMs: Long? = null,
     /** Rider's own bike speed in m/s, sourced from the radar's
      *  device-status frame (byte[len-1] x 0.25 m/s per LSB - native
-     *  protocol resolution). Null until the first device-status frame
-     *  has been received in the current session. UI consumers convert
-     *  to km/h at presentation time; HA wire format keeps km/h via a
-     *  conversion at the publish boundary. */
-    val bikeSpeedMs: Int? = null,
+     *  protocol resolution). Carried as Float so threshold comparisons
+     *  in AlertDecider / ClosePassDetector / decoder land on the exact
+     *  raw-byte boundaries the prior km/h thresholds did. Null until
+     *  the first device-status frame has been received in the current
+     *  session. */
+    val bikeSpeedMs: Float? = null,
 ) {
     val isClear: Boolean get() = vehicles.isEmpty()
 }
